Fixed the dungeon 3 boss behaviour.
The Master Arbror was not counting its sons right, as they
were removed two times when becoming vulnerable
(once with the remove_sons() call, and once with the disappear event,
leading to a "end immobilized" message.
The last immobilized son was sometimes removed after the tree became
hostile again, so the immobilized sons counter was improperly
set to -1 in that case.

The easier way I found to avoid such bugs was to check the bounds
used there (0,3) when adding or removing an immobilized son.

I also added a remove_sons() call at the boss's death event, so that
the same last immobilized son would disappear evenually, preventing
also a segmentation fault to me.

With this patch I finally was able to kill the dungeon 3 boss :)