Implement real parallax scrolling.
Now, parallax scrolling tiles are not displayed at their position
anymore, but elsewhere on the map with an offset that depends
on the camera position.
Before, it was a hack that only worked if the pattern was a repeated 8*8
square.
The new implementation is simpler but requires to edit existing maps
having parallax tiles. Parallax tiles such as holes have to be moved and
resized so that they appear where they should.