Reimplement old-style parallax tiles and call them self-scrolling tiles
Short version: what we used to call parallax scrolling before today was
not parallax scrolling but can be useful. We call this self-scrolling
tiles, and both scrolling types are available now.

Long version:
A self-scrolling tile scrolls on itself of 1 pixel when the camera4
moves of 2 pixels, achieving the same effect as parallax scrolling.
However, unlike parallax scrolling, the place
where they are displayed never changes when the camera moves.
As a consequence, they only work with repeatable 8*8 patterns.
Self-scrolling tiles are useful to implement holes since holes
must be displayed at a precise location and their graphics are a
repeated pattern of 8*8 pixels.
Real parallax scrolling is used to make nice and complex scrolling
backgrounds.