Map editor: fix an entity duplication bug.
When you create an entity that shows a dialog at its creation (some
entities like enemies and teletransporters do that because they require
the user to fill mandatory parameters), this entity does not belong to
the map yet. And if you changed its layer during this dialog, a bug
made the map think that the entity was on both layers.