Don't restore the previous hero's direction after the victory state
If there is a teletransportation after the victory state, the
destination point determines the direction, otherwise keeping north is
just fine.

This fixes a bug introduced by commit 56594268 (In the victory
state, give direction north to the sprites). When leaving dungeons, the
hero did not always face south as expected.