Timer: restore the initial suspending behavior of v0.9 and document it.
Like in Solarus 0.9, timers are now suspended at their creation only if
there is a dialog (and not if there is only a camera movement!).
After that, by default, they get suspended like the map.

This rule might seem complex, but we need it to allow quest makers to
keep their timers directly working.