Fix blocking streams not activated when they could by aligning the hero.
Streams with a narrow exit of 16 pixels were not activated unless the hero
was precisely aligned with their exit.
But they should: when the hero is not precisely aligned with
a stream's exit, the stream automatically aligns him once activated.

This made the blocks+streams puzzle of dungeon 8 1F in zsdx very easy.