Replacing ref/unref calls by the use of std::shared_ptr.
The transition will be progressive and everything should continue to
work during it.

During the transition phase, our old refcount system remains in place,
and objects managed by std::shared_ptr actually use it though a custom
deleter. This is temporary, this is only to allows both systems to
coexist during the transition.

After the transition, the old refcount system and the special deleter
will be removed. Objects will use the standard deleter of
std::shared_ptr.