Bug #492: Use of CMake variables with plural wording in the suffix
A few scripts for the command "find_package" provide the relevant
CMake variables only in the singular form while others set also variants
with a plural wording in the suffix. This fact is respected in the main
build script now.

The content of a CMake variable which contains only one list element
should be quoted in this application context so that special characters
will also be preserved.
http://cmake.org/Wiki/CMake/Language_Syntax#CMake_splits_arguments_unless_you_use_quotation_marks_or_escapes.

Signed-off-by: Markus Elfring <elfring@users.sourceforge.net>