Added license badge to the README.
I took the liberty to add a GPLv3 badge to the README: it's cute, it makes
a nice sibling to the Travis badge and it makes it clear right from the
start what the license is (which is somewhat important because of the
copyleft). It can't be bad, right? :p